requirements.txt
google-cloud-texttospeech
requests
librosa
requests-testadapter
flask
python-dotenv